Improve offline install procedure (#174)
If /QOpenSys/usr/bin is set ahead on the PATH, then the touch command
will fail as it finds the PASE command instead of the QSH command. This
was only done so that the QSH output redirect would not convert to
EBCDIC, but we can bypass all that by executing the whole command and
redirect within a PASE shell instead and we don't need the separate
touch command.