Add Zulip forum link
Years ago, Ryver started charging for their service, however we were
grandfathered in due to being an open source community. This
"grandfathered" status was ubruptly taken away around March 27th 2025.
Since then, Ryver is locked and in read-only mode.

Some members of the community surveyed alternatives and a Zulip server
was set up for us at https://chat.ibmioss.org by Calvin Buckley
(thanks!). Zulip is open source (https://github.com/zulip/zulip) and
able to be self-hosted, so we don't have to worry about the rug getting
pulled again.

This adds the new Zulip forum and marks it as the replacement for Ryver
for those looking. The Ryver forum link is also replaced from the
URL-shortened version to the full version and marked as archived. The
shortened URLs will be updated to point at Zulip in the future.